rename io_params to screen_io
