+++ commec/tests/test_screen_io.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import os
 import argparse
 
-from commec.config.io_parameters import ScreenIO
+from commec.config.screen_io import ScreenIO
 from commec.screen import add_args
 
 INPUT_QUERY = os.path.join(os.path.dirname(__file__), "test_data/single_record.fasta")
